pyproject.toml
[project]
name = "pantheon-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Jeffrey04", email = "mycyberpet@yahoo.com" }
]
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=23.1.0",
    "fastapi[standard]>=0.115.6",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "structlog>=25.1.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
pantheon-assignment = "pantheon_assignment:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
